chore: remove home page component
