+++ apps/web/vibes/soul/examples/pages/home/electric.tsx
@@ -5,9 +5,65 @@ import { featuredProducts } from '@/vibes/soul/examples/sections/featured-produc
 import { copyright, footerLinks } from '@/vibes/soul/examples/sections/footer/electric'
 import { headerLinks } from '@/vibes/soul/examples/sections/header/electric'
 import { heroSlides } from '@/vibes/soul/examples/sections/slideshow/electric'
-import { HomePage } from '@/vibes/soul/pages/home'
-import { FeaturedImageProps } from '@/vibes/soul/sections/featured-image'
+import { AnnouncementBar } from '@/vibes/soul/primitives/announcement-bar'
+import { Feature } from '@/vibes/soul/sections/feature'
+import { FeaturedCardCarousel } from '@/vibes/soul/sections/featured-card-carousel'
+import { FeaturedImage, FeaturedImageProps } from '@/vibes/soul/sections/featured-image'
+import { FeaturedProductsCarousel } from '@/vibes/soul/sections/featured-products-carousel'
+import { FeaturedProductsList } from '@/vibes/soul/sections/featured-products-list'
+import { Footer } from '@/vibes/soul/sections/footer'
+import {
+  Amex,
+  ApplePay,
+  Bitcoin,
+  GooglePay,
+  Mastercard,
+  Paypal,
+  Visa,
+} from '@/vibes/soul/sections/footer/payment-icons'
+import { Facebook, Instagram, X, Youtube } from '@/vibes/soul/sections/footer/social-icons'
+import { Header } from '@/vibes/soul/sections/header'
+import { Slideshow } from '@/vibes/soul/sections/slideshow'
+import { Subscribe } from '@/vibes/soul/sections/subscribe'
 import { ProductCardProduct } from '@/vibes/soul/types'
+
+const logo = 'SOUL'
+
+const socialMediaLinks = [
+  {
+    href: '#',
+    icon: <Facebook />,
+  },
+  {
+    href: '#',
+    icon: <X />,
+  },
+  {
+    href: '#',
+    icon: <Instagram />,
+  },
+  {
+    href: '#',
+    icon: <Youtube />,
+  },
+]
+
+const paymentIconsArray: React.ReactNode[] = [
+  <Visa key="Visa" />,
+  <Amex key="Amex" />,
+  <Mastercard key="Mastercard" />,
+  <Paypal key="Paypal" />,
+  <GooglePay key="GooglePay" />,
+  <ApplePay key="ApplePay" />,
+  <Bitcoin key="Bitcoin" />,
+]
+
+const locales = [
+  { id: '1', region: 'US', language: 'EN' },
+  { id: '2', region: 'FR', language: 'FR' },
+  { id: '3', region: 'DE', language: 'DC' },
+  { id: '4', region: 'IT', language: 'IT' },
+]
 
 // Featured Products
 export const newArrivals: ProductCardProduct[] = [
@@ -137,18 +193,81 @@ export const featuredImageII: FeaturedImageProps = {
 
 export default function Preview() {
   return (
-    <HomePage
-      headerLinks={headerLinks}
-      logo="SOUL"
-      heroSlides={heroSlides}
-      categories={cards}
-      featuredImage={featuredImage}
-      feature={feature}
-      featuredProducts={featuredProducts}
-      newArrivals={newArrivals}
-      featuredImageII={featuredImageII}
-      footerLinks={footerLinks}
-      copyright={copyright}
-    />
+    <>
+      <AnnouncementBar>
+        Get <strong>15% off</strong> and free shipping with discount code{' '}
+        <strong>&quot;welcome&quot;</strong>
+      </AnnouncementBar>
+      <Header
+        links={headerLinks}
+        logo={logo}
+        cartHref="#"
+        accountHref="#"
+        locales={locales}
+        activeLocale="EN"
+      />
+      <Slideshow slides={heroSlides} />
+      <FeaturedCardCarousel title="Categories" cards={cards} />
+      <FeaturedImage
+        title={featuredImage.title}
+        description={featuredImage.description}
+        image={{
+          src: featuredImage.image.src,
+          alt: featuredImage.image.alt,
+        }}
+        cta={{ href: '#', label: 'Shop Now' }}
+      />
+      <FeaturedProductsList
+        title={featuredProducts.title}
+        description={featuredProducts.description}
+        cta={{
+          label: featuredProducts.cta?.label ?? '',
+          href: featuredProducts.cta?.href ?? '',
+        }}
+        products={featuredProducts.products}
+      />
+      <Subscribe
+        title="Sign up for our newsletter"
+        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor."
+      />
+      <FeaturedProductsCarousel
+        title="New Arrivals"
+        cta={{ label: 'See All', href: '#' }}
+        products={newArrivals ?? featuredProducts.products}
+      />
+      <Feature
+        image={{
+          src: feature.image.src,
+          alt: feature.image.alt,
+        }}
+        title={feature.title}
+        description={feature.description}
+        cta={{
+          label: feature.cta.label,
+          href: feature.cta.href,
+        }}
+      />
+      <FeaturedProductsCarousel title="Recently Viewed" products={featuredProducts.products} />
+      <FeaturedImage
+        title={featuredImageII.title}
+        description={featuredImageII.description}
+        image={{
+          src: featuredImageII.image.src,
+          alt: featuredImageII.image.alt,
+        }}
+        cta={{ href: featuredImageII.cta.href, label: featuredImageII.cta.label }}
+      />
+      <Subscribe
+        title="Sign up for our newsletter"
+        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor."
+      />
+      <Footer
+        logo={logo}
+        sections={footerLinks}
+        copyright={copyright}
+        paymentIcons={paymentIconsArray}
+        socialMediaLinks={socialMediaLinks}
+      />
+    </>
   )
 }
